+++ apps/v4/components/component-preview.tsx
@@ -10,6 +10,7 @@ export function ComponentPreview({
   className,
   align = "center",
   hideCode = false,
+  chromeLessOnMobile = false,
   ...props
 }: React.ComponentProps<"div"> & {
   name: string
@@ -17,12 +18,13 @@ export function ComponentPreview({
   description?: string
   hideCode?: boolean
   type?: "block" | "component" | "example"
+  chromeLessOnMobile?: boolean
 }) {
   const Component = Index[name]?.component
 
   if (!Component) {
     return (
-      <p className="text-muted-foreground text-sm">
+      <p className="text-muted-foreground mt-6 text-sm">
         Component{" "}
         <code className="bg-muted relative rounded px-[0.3rem] py-[0.2rem] font-mono text-sm">
           {name}
@@ -63,6 +65,7 @@ export function ComponentPreview({
       hideCode={hideCode}
       component={<Component />}
       source={<ComponentSource name={name} collapsible={false} />}
+      chromeLessOnMobile={chromeLessOnMobile}
       {...props}
     />
   )