+++ apps/v4/registry/new-york-v4/ui/field.tsx
@@ -196,12 +196,12 @@ function FieldError({
       return children
     }
 
-    if (!errors) {
+    if (!errors?.length) {
       return null
     }
 
-    if (errors?.length === 1 && errors[0]?.message) {
-      return errors[0].message
+    if (errors?.length == 1) {
+      return errors[0]?.message
     }
 
     return (